# Optimise buffer usage in zip walk functions