--- a/changelog/@unreleased/pr-94.v2.yml
+++ b/changelog/@unreleased/pr-94.v2.yml
@@ -1,0 +1,6 @@
+type: improvement
+improvement:
+  description: Some optimisations have been made which should reduce the resource
+    overhead required when inspecting zip files (i.e. jars, wars, pars, etc.)
+  links:
+  - https://github.com/palantir/log4j-sniffer/pull/94